requirements.txt
torch>=2.0.0
numpy
pandas
tqdm
pyyaml
prettytable
scikit-learn
matplotlib
seaborn
jupyter
